move iface: further trajectory checks and global action client
Check current state for moving to home and for pick_place. Only the gripper
in simulation is allowed to have a goal state close to its current state.
This fix also includes a method to check cartesian poses for closeness in
joint space. The distance check is increased to 0.01 because 0.005 was not
enough in some cases when testing with the real robot and the Path Planning
project.

The action client is now running all the time and not initialized in the action
server callback. This allows to call the action client multiple times without
having to restart the interface client. The disadvantace is a warning message
when restarting the interface client (reregistering of goal/cancel handle).
I tried several other possibilities of handling this and none of them worked.
Reinitializing the action client outside the callback via global trigger
variables had the same effect as doing it in the server callback. Maybe there
is some issue with the async spinner and intializing of action clients after
the spinner was started.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>