Add move_client package
This package aims to help users interface with the RLLMoveIface. It provides methods to call services
provided by a move interface, see: #22

The MoveClient is designed to be extensible and applicable for any type of move interface. To allow this,
the logic for the different movement services is split into multiple classes, i.e. for basic and pick 'n'
place movements (and could(/should?) be split even further). This makes it easy to create a specifically
tailored MoveClient* for any kind of project by combining the corresponding classes. The MoveClient logs
all requests and responses by default and validates the received responses. If a response indicates a
critical error, an exception is raised. If a non critical failure occurs the service call success status
is set to false and no exception is raised. If desired, the MoveClient can be configured to always raise
an exception on failure, which can make error validation simpler.

There are two approaches on how to use the MoveClient to execute user code. The first possibility is to
derive and override the execute() method. The second option is to specify a callback function, which might
be the simpler option for beginners.

The client is available for python and C++. See the move_client_example.{py, cpp} for an example on how to
use the library.

The code is formatted with clang-format, flake8. And mostly clang-tidied (only ROS internal/related such
as in ROS_INFO_STREAM or messages with bools).