Document the Robot Playground hello_world example
Document the usage of the move_client functions for Python and C++. The documentation uses the copied files, included with literalinclude, from the rll_robot_playground_project. This has the drawback, that the files are duplicated, where are relative symlink might have also worked. However, if the files are changed in the rll_robot_playground_project the reffered to lines in literalinclude might/will be wrong and its better to work on a copy and only update the files if you adjust the literalincludes. To make dealing with the copied files easier two new make rules have been added.

To display C++ and Python side by side the the sphinx.tabs extension is used. For some reason the stylesheets of sphinx.tabs are not added automatically, so I added them manually.