Move iface startup wait
The move_iface requires Moveit to be setup before it can be started.
Currently the user has to take care of this themselves, see #45.
To overcome this limitation, the waitForMoveGroupAction() function is introduced,
which checks (and waits for a time) if the `move_group` action is present.

If you call waitForMoveGroupAction() in your ROS node code before creating the
move_iface instance, you can start the move_iface and Moveit setup simultaneously.
To showcase this, the moveit_planning_execution_and_move_iface.launch script has been added.

Note: I moved initConstTransforms() into the RLLMoveIface constructor's startup checks
and removed it from RLLDefaultMoveIfaceBase because this check should be always done.
By putting it here, the check doesn't have to be added in every inherited interface.