Reworked move interface with new TCP listener for the client
So far the move interface client was run as an action server. However this setup proved to be unreliable
because goals get lost in a short time window when a new action server needs to be registered in the
action client. This merge request replaces the action server setup for the client with a simple TCP
listener socket that exchanges predefined messages with the client. The main purpose is to notify the
client that it can start running its main function.

The move client library is adapted accordingly. I reworked the move interface itself to improve code
structure. The original general move interface class is split up into a class for the basic planning
methods and a dedicated class for the services the interface provides. For tracking the job result, there
is now a job_result class that can be used to set the job result, either by the client through a service
or by the interface itself.

The whole SDK is reworked using pylint, catkin_lint and version 6.0 of clang-tidy and clang-format. The CI
is extended to run these tools automatically. Furthermore, warnings are enabled for catkin build.

See the no_client_action_srv branch for more details.