move iface: several bugfixes
Permissions are restored in runJob() because they might not
be restored in runClient() if runClient() returns early.

Internal Error state of the interface is checked after it is
verified that there is not service running. This ensures that
also an internal error during the last service run is catched.

strncmp is used instead of strcmp for better handling of not
properly terminated strings.

Thanks to Mark Weinreuter for pointing out these bugs in !46.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>