move iface: always assume current pose as start pose in poseGoalTooClose()
Every use case of this method right now passes the current pose as start
pose anyway. And in some rare cases (singularities) the IK from the current
pose fails, which leads to this check failing although it probably should
succeed (as there always should be a IK solution to the current robot pose).

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>