MoveClient ensure job_finished is called
Currently, if the move_client execution fails or is interrupted by the user (SIGINT),
the move_iface is not always notified about the job result or the jobFinished
service call fails due to concurrent service calls. In this case the move_interface
continues to wait until the job execution timeout has expired. This is cumbersome if
you want to repeatedly test your client code and have to restart the move_iface to
accept a new job. To circumvent this the move_client now calls the jobFinished
service even if it is interrupt or an unhanded exception occurs in the client code.
Furthermore the jobFinished service still sets the job result (to false) even if it
is a concurrent service call, which would otherwise be ignored.

Additionally, some new utility functions are added to the move_client, however, at
this time for Python only.