Reworked Kinematics package and refactored move interface
The whole kinematics package is renamed to the more general name
rll_kinematics and unneeded dependencies are removed.

In general, the kinematics package has now a more modular design with
lots of duplicate code removed. More C++ features and static_vector
from Boost are used. There should not be any dynamic memory allocation
inside the IK any more.

The kinematics code is also more efficient now. All coefficients for IK
formulas are collected and partly precalculated to avoid costly
recalculations. There is also a single interface for computation of
joint and arm angles, derivatives and the mapping between joint space
and null space.

Several issues from the previous implementation are fixed. This mostly
affects the selection of feasible intervals and closest configs, and the
mapping for overlapping arm angle intervals. Singularities and joint
limits are also checked for the forward kinematics.

Several helper methods are used to avoid numerical inaccuracies,
especially regarding zero rounding, square root, acos() and joint limits.
Additionally, several useful custom data types are added.

The code is split up in the different components of the IK solution:
forward kinematics, general inverse kinematics, computation of
feasible intervals and redundancy resolution. If one is familiar with
the related papers, it should now be easy to navigate the code.

The API for the inverse kinematics is completely reworked and should
now be easy to use. It is now possible to select the resolution method
and specify parameters for the resolution during runtime. Error codes
similar to the ones used for the move interface are implemented which
do now a clear distinction between success (includes warnings) and
errors. An example is added to show how the API can be used (besides
using the Moveit Plugin).

In the move interface and the kinematics plugin, the planning with
desired arm angles is improved and support for these commands is
addded to the move client. This also includes the ability to request
the current arm angle and global configuration.

In a lot of places in the codebase, the name "elbow angle" is used
instead of the proper name arm angle. This is fixed now.

Several tests are added to the move interface: tests for movement
sequences (similar to hello worlds, demos), invalid movements, tests
for planning with desired arm angle.

An important fix for the long-standing bug with invalid start states
(deviations) is also included. It is now verified that the robot state
in the Moveit planning scene has the final goal pose. This way, a race
condition is avoided when the next planning request is processed
before the robot state was updated. A common execute() method is now
used to ensure that this check is done for every motion request.
The fix allows to switch back to the nicer interpolating fake
controller and to enforce a stricter start state tolerance.

The time-optimal path parameterization is now used by default on
Melodic which leads to nicer trajectories in simulation.

A custom bool service return types is defined to avoid implicit bool
conversions.

Initial work to refactor the planning class is included. More methods
and class members are now private and references and pointers are
properly used for method parameter, except for service definitions
where this is not possible.

The pedantic compiler flag is enabled for all C++ packages to ensure
C++ ISO standard compliance. Clang-tidy is updated to version 7 and
the config is improved. There are limits for function/method length
and number of parameters. The return code of the clang-tidy script
now works in the CI system, but warnings need to be treated as
errors for this to be effective.