Kinematics fixes and unified build options
Several issues with the inverse kinematics are fixed:
* All possible cases in the arm angle interval search should now be
  covered and I did not get any exceptions anymore.
* The singularity check, that checks if axis 1 and 7 are aligned, did not
  check if the seventh axis is vertical, only if the wrist is above the
  shoulder.
* A fallback is added if the elbow angle is zero. Then the global
  configuration at the joint is not uniquely defined, so arm angles for both
  configs need to be checked. The arm angle is actually not defined with this
  joint angle value, but the formulas are still able to return feasible
  solutions, albeit not optimal.
* The redundancy resolution with the exponential function did not compute the
  joint angles if there was only one possible arm angle. And the closest
  config search did not return the arm angle and global config from the
  selected solution, but from the latest computed one.
* Floating point values are not directly compared anywhere anymore. Instead
  several new helper methods are used and the zero rounding tolerance is
  enforced in more places.
* Hinge joint singularities need to be treated to avoid the creation of wrong
  blocked intervals.
* The margin around pivot singularities for blocked intervals was way too big
  and is now set to a sensible value.
* The arm angle is directly set to zero in the fallback case (no feasible
  interval found) to ensure that a sensible arm angle value is used. Zero
  should correspond to the reduced 6-DOF robot arm and the joint values are
  feasible, but the values are often strange in singularities so the behavior
  of the formulas in these cases needs to be checked. And it could be
  necessary to switch to an IK algorithm for 6-DOF robots to achieve stable
  behavior in singularities.

The inverse kinematics was tested with 1 billion feasible poses and did
return a solution within the joint limits for all poses. The code was
optimized in a lot of places, primarily to reduce calls to costly
trigonometric functions and to speed up matrix operations. I did an initial
benchmark and the average execution time is around 7 microseconds, while
the maximum execution time is below 40 microseconds.

The kinematics package is split up into a package with the library and a
separate package with the Moveit plugin. Then it is possible to directly
use the library without Moveit dependencies.

Among the tests, the kinematics test E1 is replaced with a new one. E1 was
a test from Philipp's evaluation. However he didn't use the gripper TCP,
but the flange as TCP. That's why the rotation is much bigger compared to
his evaluation. And when rotating around the gripper TCP, a joint limit in
axis 2 is either touched or crossed. This is the reason why the test failed
sometimes.
The new test case is an interesting rotation around the robots base that
could be useful for the evaluation of the IK.

A fix for the move interface is also included: The return code of
computeLinearPathArmangle() was not checked in the service call.

Cmake build options are now collected in the rll_core package and are
used automatically in all packages that either depend on rll_core directly
or on rll_msgs and rll_kinematics. The release build type is now set by
default which enables O3 optimizations. If you want debugging symbols, you
now need to specify the debug build, e.g. via catkin build on the command
line: catkin build --cmake-args -DCMAKE_BUILD_TYPE=Debug -- rll_kinematics.
More warnings are also enabled.

The linting commands for the CI are collected in a script because the single
command line in the gitlab-ci.yaml became too big.