Check all global configurations for PTP movements and Clang-Tidy 9 fixes for IK
Now all global configurations are checked for PTP movements. The IK is able to return
multiple solutions, each one is for a different global configuration. This update
primarily improves the behavior of the collision check for PTP movements. If one
configuration is in collision, a different one can now be selected that is not in
collision. The joint values from the selected solution are afterwards used for
planning.

However this is not a proper redundancy resolution with collision check. I might add
something like this in the future. It is currently still possible that an arm angle
is selected in all global configurations that leads to a collision, although an arm
angle exists that would not result in a collision. The arm angle is selected depending
on the seed state and the selected arm angle should avoid jumps in joint space, but
the redundancy resolution does not know about any collision objects for which the arm
angle might be unsuitable.

The IK is generally now able to return multiple solutions. In the default mode, the IK
returns solutions for all configurations that were checked. These are only the ones
that are close to the seed state. With a new flag, it is now also possible to return
solutions for all possible global configurations.

The IK is also updated to check for invalid input (NaN, infinite values) or calls to FK
or IK methods without proper initialization of the robot properties.

Several linting fixes are included to make clang-tidy 9 happy. Interval data members are
now private and more methods are static. A few of the new checks are disabled in the
clang-tidy config because they either cause false positives or because they are generally
not desirable. Using trailing return type for all method declarations does not improve
readability. Putting every variable declaration on a separate line is unnecessary if they
have the same type and are not initialized. The magic numbers check cannot be enforced in
our case, this is something that needs to be checked in code reviews.