description: various fixes for dynamics properties and controller settings
The inertia parameters for the EGL gripper are updated with the values
from the identification by the Sunrise controller. The real joint torque
limits are added for the iiwa7. The table inertias are set to more
realistic values.

The legacy mode tag is not needed anymore for the Gazebo ROS control
interface. The Gazebo grasp fix plugin settings are updated to make it
attach grasp objects right away on first contact while keeping the update
rate low.
The iiwa actuator hardware interfaces need to be prepended with
"hardware_interface". The safety controller pos and vel factors are set to
more sensible values.

The surface property settings are cleaned up. Unless more fine-grained
settings are needed in the future, it is enough to set mu1 and mu2.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>