description: use a simplified collision model for iiwa
The model is similar to the collision model of the Sunrise controller
which exclusively uses spheres. The Sunrise controller model is detailed in
the Sunrise OS docs.
However the Sunrise controller model does not cover the whole robot. Parts
of link 1, 3 and 5 are not covered by the small spheres.
To overcome this limitation, cylinders are used for these three links. The
cylinder dimensions were chosen to cover the whole link while not being
smaller than the spheres from the Sunrise model at any point.

This new collision model should make it easier to be compatible with the
Sunrise model and to avoid discrepancies between the two when running
trajectories on the real robot.
The model also leads to faster collision checking in Moveit and reduces
CPU load in Gazebo. The accuracy of the model should still be sufficient
for our use cases and it should not limit motions that were previously
possible.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>