description: tweak Gazebo grasp plugin settings
The larger release tolerance helps to place objects in a more stable
state and ensures that grasp objects are not quickly detached and
attached again when placing objects.

The increased update rate seemed to make the attaching more stable and
I observed less jumping of grasp objects into the grasp position.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>