New redundancy resolution and related Move Interface extensions
This mainly introduces the new multi-objective redundancy resolution for the IK. This method is now also used by default. Small fixes for the other IK modes and the fallback arm angle selection are also included.

New pass-through methods for the kinematics plugin are used to set IK options. The IK needs velocity and acceleration limits, which are set from the configured limits in the Moveit config. Multiple seed states are also needed for the new redundancy resolution.

Regarding the move interface: A custom computeCartesianPath() is now used, which allows us to set IK options and influence the sampling. The interpolation for linear movements is reworked: Sole rotations are now also supported. Arm angles are properly interpolated for moveLinArmangle.

The tests are adapted accordingly. Small fixes were needed to account for differences between the old and the new IK.

I broke the getCurrentPose service in my last merge request. The service works now again. The collision check for the goal in the moveLin service is moved to moveToGoalLinear() to ensure that it is always called. I also added a FK pose to world helper, which does the inverse of the pose transform helper for the IK.